+++ packages/@baldeweg/ui/src/components/BAlert/BAlert.stories.js
@@ -8,8 +8,12 @@ export default {
       control: { type: 'select', options: ['neutral', 'info', 'danger', 'warning', 'success'] },
     },
     closable: { control: 'boolean' },
+    position: {
+      control: { type: 'select', options: ['block', 'top', 'bottom', 'topleft', 'topright', 'bottomleft', 'bottomright'] },
+      defaultValue: 'block',
+    },
   },
-  decorators: [() => ({ template: '<div style="padding: 3em; height: 300px; position: relative;"><story/></div>' })],
+  decorators: [() => ({ template: '<div style="padding: 3em; height: 300px; position: relative;">This is same text.<story/></div>' })],
 
 };
 
@@ -51,4 +55,51 @@ Success.args = {
   closable: true,
 };
 
+export const Block = Template.bind({});
+Block.args = {
+  type: 'neutral',
+  closable: false,
+  position: 'block',
+};
 
+export const Top = Template.bind({});
+Top.args = {
+  type: 'info',
+  closable: true,
+  position: 'top',
+};
+
+export const Bottom = Template.bind({});
+Bottom.args = {
+  type: 'success',
+  closable: true,
+  position: 'bottom',
+};
+
+export const TopLeft = Template.bind({});
+TopLeft.args = {
+  type: 'warning',
+  closable: false,
+  position: 'topleft',
+};
+
+export const TopRight = Template.bind({});
+TopRight.args = {
+  type: 'danger',
+  closable: true,
+  position: 'topright',
+};
+
+export const BottomLeft = Template.bind({});
+BottomLeft.args = {
+  type: 'warning',
+  closable: false,
+  position: 'bottomleft',
+};
+
+export const BottomRight = Template.bind({});
+BottomRight.args = {
+  type: 'danger',
+  closable: true,
+  position: 'bottomright',
+};